+++ CHANGELOG.md
@@ -6,12 +6,17 @@
     Click to see more.
   </summary>
 
+### Major
+* GroupAvatar: `size` prop is now optional. (#244)
+
 ### Minor
 * Masonry: Makes Masonry React Async compatible (#227)
 * SegmentedControl: Change flow type of `items` to `React.Node` (#230)
 * Video: Add jsdom browser specific tests (#205)
 * Flyout: Merge ErrorFlyout abilities into Flyout (#242)
 * Card: Make Card explicitely use box-sizing: content-box (#243)
+* GroupAvatar: Text sizes have shrunk to be consistent with Avatar. (#244)
+* Video: Move initial video setup calls to componentDidMount (#245)
 
 ### Patch
 * Internal: add better basic test coverage (#231)